+++ mex/common/models/activity.py
@@ -11,6 +11,7 @@ from mex.common.models.base.preview_item import PreviewItem
 from mex.common.models.base.rules import (
     AdditiveRule,
     PreventiveRule,
+    PublishingRule,
     RuleSet,
     SubtractiveRule,
 )
@@ -26,6 +27,7 @@ from mex.common.types import (
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
     MergedPrimarySourceIdentifier,
+    PublishingStatus,
     Text,
     Theme,
     Year,
@@ -426,12 +428,25 @@ class PreventiveActivity(_Stem, PreventiveRule):
     website: list[MergedPrimarySourceIdentifier] = []
 
 
+class PublishingActivity(_Stem, PublishingRule):
+    """Rule to prevent publishing of merged activity items."""
+
+    entityType: Annotated[
+        Literal["PublishingActivity"], Field(alias="$type", frozen=True)
+    ] = "PublishingActivity"
+    status: Annotated[
+        PublishingStatus | None,
+        Field(description="Indicates if the merged item should NOT be published."),
+    ] = None
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for an activity item."""
 
     additive: AdditiveActivity = AdditiveActivity()
     subtractive: SubtractiveActivity = SubtractiveActivity()
     preventive: PreventiveActivity = PreventiveActivity()
+    publishing: PublishingActivity = PublishingActivity()
 
 
 class ActivityRuleSetRequest(_BaseRuleSet):