+++ mex/common/models/person.py
@@ -11,6 +11,7 @@ from mex.common.models.base.preview_item import PreviewItem
 from mex.common.models.base.rules import (
     AdditiveRule,
     PreventiveRule,
+    PublishingRule,
     RuleSet,
     SubtractiveRule,
 )
@@ -20,6 +21,7 @@ from mex.common.types import (
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
     MergedPrimarySourceIdentifier,
+    PublishingStatus,
 )
 
 EmailStr = Annotated[
@@ -339,12 +341,25 @@ class PreventivePerson(_Stem, PreventiveRule):
     orcidId: list[MergedPrimarySourceIdentifier] = []
 
 
+class PublishingPerson(_Stem, PublishingRule):
+    """Rule to prevent publishing of merged person items."""
+
+    entityType: Annotated[
+        Literal["PublishingPerson"], Field(alias="$type", frozen=True)
+    ] = "PublishingPerson"
+    status: Annotated[
+        PublishingStatus | None,
+        Field(description="Indicates if the merged item should NOT be published."),
+    ] = None
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for a person item."""
 
     additive: AdditivePerson = AdditivePerson()
     subtractive: SubtractivePerson = SubtractivePerson()
     preventive: PreventivePerson = PreventivePerson()
+    publishing: PublishingPerson = PublishingPerson()
 
 
 class PersonRuleSetRequest(_BaseRuleSet):